% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ClaudeModel.R
\name{ClaudeModel}
\alias{ClaudeModel}
\title{Generate a Claude response, optionally including submission and solution images.}
\description{
Generate a Claude response, optionally including submission and solution images.

Generate a Claude response, optionally including submission and solution images.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-ClaudeModel-new}{\code{ClaudeModel$new()}}
\item \href{#method-ClaudeModel-generate_response}{\code{ClaudeModel$generate_response()}}
\item \href{#method-ClaudeModel-clone}{\code{ClaudeModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClaudeModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-ClaudeModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClaudeModel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClaudeModel-generate_response"></a>}}
\if{latex}{\out{\hypertarget{method-ClaudeModel-generate_response}{}}}
\subsection{Method \code{generate_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClaudeModel$generate_response(
  prompt,
  system_instructions,
  submission_images = NULL,
  solution_image = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-ClaudeModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-ClaudeModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ClaudeModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
