% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_utils.R
\name{extract_question_from_txt}
\alias{extract_question_from_txt}
\title{Extract a specific question block from a submission file}
\usage{
extract_question_from_txt(submission, question)
}
\arguments{
\item{submission}{Path to the submission file}

\item{question}{The exact heading string to look for}
}
\value{
Text block belonging to the specified question
}
\description{
Extract a specific question block from a submission file
}
