% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_utils.R
\name{gather_file_contents}
\alias{gather_file_contents}
\title{Generate file contents with line numbers}
\usage{
gather_file_contents(file_paths, question = NULL)
}
\arguments{
\item{file_paths}{List of file paths to process}

\item{question}{Optional question heading to extract from files}
}
\value{
Character string with file contents and line numbers
}
\description{
Generate file contents with line numbers
}
