% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/template_utils.R
\name{gather_file_references}
\alias{gather_file_references}
\title{Generate file reference descriptions}
\usage{
gather_file_references(submission, solution = NULL, test_output = NULL)
}
\arguments{
\item{submission}{Path to student's submission file}

\item{solution}{Path to instructor's solution file (optional)}

\item{test_output}{Path to test output file (optional)}
}
\value{
Character string with file descriptions
}
\description{
Generate file reference descriptions
}
