% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{main}
\alias{main}
\title{AI feedback entrypoint}
\usage{
main(
  prompt = NULL,
  prompt_text = NULL,
  prompt_custom = NULL,
  scope,
  submission,
  solution = "",
  model,
  remote_model,
  output = "",
  submission_image = NULL,
  solution_image = NULL,
  output_template = NULL,
  system_prompt = NULL,
  question = NULL,
  marking_instructions = NULL
)
}
\description{
AI feedback entrypoint
}
