% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OpenAIModel.R
\name{OpenAIModel}
\alias{OpenAIModel}
\title{Initialize the OpenAIModel instance by loading the API key.
Generate a model response using the OpenAI API, including optional images.}
\description{
Initialize the OpenAIModel instance by loading the API key.
Generate a model response using the OpenAI API, including optional images.

Initialize the OpenAIModel instance by loading the API key.
Generate a model response using the OpenAI API, including optional images.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-OpenAIModel-new}{\code{OpenAIModel$new()}}
\item \href{#method-OpenAIModel-generate_response}{\code{OpenAIModel$generate_response()}}
\item \href{#method-OpenAIModel-clone}{\code{OpenAIModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAIModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAIModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAIModel$new()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAIModel-generate_response"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAIModel-generate_response}{}}}
\subsection{Method \code{generate_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAIModel$generate_response(
  prompt,
  system_instructions,
  submission_images = NULL,
  solution_image = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-OpenAIModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-OpenAIModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{OpenAIModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
