% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RemoteModel.R
\name{RemoteModel}
\alias{RemoteModel}
\title{Initialize RemoteModel with a remote URL and model name
Generate a model response using the prompt and optional file paths.}
\description{
Initialize RemoteModel with a remote URL and model name
Generate a model response using the prompt and optional file paths.

Initialize RemoteModel with a remote URL and model name
Generate a model response using the prompt and optional file paths.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-RemoteModel-new}{\code{RemoteModel$new()}}
\item \href{#method-RemoteModel-generate_response}{\code{RemoteModel$generate_response()}}
\item \href{#method-RemoteModel-clone}{\code{RemoteModel$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteModel-new"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteModel-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteModel$new(
  remote_url = "http://polymouth.teach.cs.toronto.edu:5000/chat",
  model_name = "deepseek-coder-v2:latest"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteModel-generate_response"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteModel-generate_response}{}}}
\subsection{Method \code{generate_response()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteModel$generate_response(
  prompt,
  system_instructions,
  submission_images = NULL,
  solution_image = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-RemoteModel-clone"></a>}}
\if{latex}{\out{\hypertarget{method-RemoteModel-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{RemoteModel$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
